pyproject.toml
[tool.poetry]
name = "code-names-bot"
version = "0.1.0"
description = ""
authors = ["Nalu Zou"]
readme = "README.md"
packages = [{include = "code_names_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0.1"
openai = "^0.28.0"

[tool.poetry.scripts]
create-scenarios = "code_names_bot.create_scenarios:main"
generate-clues = "code_names_bot.generate_clues:main"
guess-clues = "code_names_bot.guess_clues:main"
evaluate = "code_names_bot.evaluate:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
